requirements.txt
Babel==1.3
Django==1.8
amqp==1.4.6
analytics-python==1.0.3
anyjson==0.3.3
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
billiard==3.3.0.20
celery==3.1.17
certifi==14.05.14
cffi==0.9.2
coverage==3.7.1
cryptography==1.3.1
customerio==0.1.6
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.16
django-celery-with-redis==3.0
django-coverage==1.2.4
django-extensions==1.5.2
django-filter==0.9.2
django-tokenapi==0.2.2
django-toolbelt==0.0.1
djangorestframework==3.1.1
enum34==1.0.4
extras==0.0.3
flower==0.8.1
futures==2.2.0
fuzzywuzzy==0.5.0
gunicorn==19.3.0
httpretty==0.8.8
ipdb==0.8
ipython==3.1.0
jdcal==1.2
jsonfield==1.0.3
kombu==3.0.24
lettuce==0.2.20
linecache2==1.0.0
mock==1.0.1
newrelic==2.50.0.39
nose==1.3.6
openpyxl==2.2.2
psycopg2==2.6
pyOpenSSL==0.15.1
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
pystache==0.5.4
python-Levenshtein==0.12.0
python-dateutil==2.4.2
python-mimeparse==0.1.4
python-subunit==1.1.0
pytz==2015.2
raven==5.2.0
redis==2.10.3
requests==2.6.0
six==1.9.0
static==1.1.1
static3==0.5.1
suds-jurko==0.6
sure==1.2.10
testtools==1.7.1
tornado==4.1
traceback2==1.4.0
uWSGI==2.0.10
unittest2==1.0.1
wsgiref==0.1.2
